Process user's chat history

+++ Tlv.Search.Common/SearchItem.cs
@@ -7,6 +7,6 @@
         public string? title { get; set; }
         public string? imageUrl { get; set; }
         public string? url { get; set; }
-        public float similarity { get; set; }
+        public double similarity { get; set; }
     }
 }
